BBB-1158 Use BigDecimal for badge cost

+++ src/main/java/uk/gov/dft/bluebadge/service/payment/client/govpay/CreatePaymentRequest.java
@@ -1,6 +1,7 @@
 package uk.gov.dft.bluebadge.service.payment.client.govpay;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.math.BigDecimal;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NonNull;
@@ -8,7 +9,7 @@ import lombok.NonNull;
 @Builder
 @Getter
 public class CreatePaymentRequest {
-  @NonNull private final Long amount;
+  @NonNull private final BigDecimal amount;
   @NonNull private final String reference;
   @NonNull private final String description;
   @NonNull private final String language;
