+++ src/main/java/uk/gov/dft/bluebadge/service/payment/repository/domain/PaymentEntity.java
@@ -1,9 +1,9 @@
 package uk.gov.dft.bluebadge.service.payment.repository.domain;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 import lombok.Builder;
 import lombok.Data;
-
 import lombok.NonNull;
 
 @Builder
@@ -13,5 +13,5 @@ public class PaymentEntity {
   @NonNull private String paymentId;
   @NonNull private String laShortCode;
   @NonNull private String reference;
-  @NonNull private Long cost;
+  @NonNull private BigDecimal cost;
 }
